requirements.txt
# ==================== Core Testing Frameworks ====================
pytest==7.4.3
pytest-playwright==0.4.3
playwright==1.40.0
pytest-xdist==3.5.0
pytest-html==4.1.1
behave==1.2.6

# ==================== Reporting ====================
allure-pytest==2.13.2
allure-behave==2.13.2

# ==================== API Testing ====================
requests==2.31.0
jsonschema==4.20.0
jsonpath-ng==1.6.1
jsonpatch==1.33
jsondiff==2.0.0

# ==================== Data Handling ====================
openpyxl==3.1.2
pandas==2.1.4
pyyaml==6.0.1
xmltodict==0.13.0

# ==================== Data Models & Validation ====================
pydantic==2.5.2
faker==20.1.0

# ==================== Configuration ====================
python-dotenv==1.0.0
configparser==6.0.0

# ==================== Logging ====================
colorlog==6.8.0

# ==================== Utilities ====================
python-dateutil==2.8.2
Pillow==10.1.0